segments are properly used
per #46

Adding functionality for re-selecting segments after user selection.

Last update changed the way interpolated data are saved in memory so they had to be dealt with differently.

There is still the question of how much of the interpolated data to show in the velocity, as of now they are all being shown